requirements.txt
fastapi
uvicorn
motor
python-jose
passlib[bcrypt]
pydantic
python-multipart
dotenv